+++ capstone/src/main/java/com/techelevator/Items/Chips.java
@@ -1,13 +1,15 @@
 package com.techelevator.Items;
 
+import java.math.BigDecimal;
+
 public class Chips extends Item{
 
     private String name;
-    private int price;
+    private BigDecimal price;
     private int quantity;
 
     //constructor
-    public Chips(String name, int price){
+    public Chips(String name, BigDecimal price){
         this.name = name;
         this.price = price;
         this.quantity = 5;
@@ -21,7 +23,7 @@ public class Chips extends Item{
     }
 
     @Override
-    public int getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 }
