+++ capstone/src/main/java/com/techelevator/Items/Gum.java
@@ -1,27 +1,28 @@
 package com.techelevator.Items;
 
+import java.math.BigDecimal;
+
 public class Gum extends Item{
 
     private String name;
-    private int price;
+    private BigDecimal price;
     private int quantity;
 
     //constructor
-    public Gum(String name, int price){
+    public Gum(String name, BigDecimal price){
         this.name = name;
         this.price = price;
         this.quantity = 5;
     }
 
     //getters
-
     @Override
     public String getName() {
         return name;
     }
 
     @Override
-    public int getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
